refactor(api): extract planning preferences service
